Final Project IR/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
argcomplete==1.10.3
asyncio==3.4.3
attrs==23.2.0
Beaker==1.12.1
beautifulsoup4==4.8.2
blinker==1.8.2
blivet==3.9.1
blivet-gui==2.5.0
boto3==1.34.51
botocore==1.34.51
Brlapi==0.8.5
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
compressed_rtf==1.0.6
cryptography==41.0.7
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
distro==1.9.0
dnf==4.19.0
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
Faker==26.0.0
fedora-third-party==0.10
file-magic==0.4.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
fros==1.1
frozenlist==1.4.1
greenlet==3.0.3
humanize==3.13.1
idna==3.6
IMAPClient==2.1.0
iso639==0.1.4
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
langtable==0.0.65
libcomps==0.1.20
libdnf==0.73.0
louis==3.28.0
lxml==5.1.0
Mako==1.2.3
MarkupSafe==2.1.3
multidict==6.0.5
mysql-connector-python==9.0.0
mysqlclient==2.2.4
nftables==0.1
olefile==0.47
packaging==23.2
Paste==3.7.1
pdfminer.six==20191110
pexpect==4.9.0
pid==2.2.3
pillow==10.2.0
ply==3.11
powerline-status==2.7
productmd==1.38
ptyprocess==0.7.0
pwquality==1.4.5
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.20.0
pycups==2.0.1
pyenchant==3.2.2
PyGObject==3.48.1
pykickstart==3.52
pyOpenSSL==23.2.0
pyparted==3.13.0
PyPDF2==3.0.1
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-docx==1.1.2
python-gettext==4.0
python-manatools==0.0.4
python-meh==0.51
python-pptx==0.6.23
pyudev==0.24.1
pyxdg==0.27
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
requests-ftp==0.3.1
rpm==4.19.1.1
s3transfer==0.10.1
selinux @ file:///builddir/build/BUILD/libselinux-3.6/src
sepolicy @ file:///builddir/build/BUILD/selinux-3.6/python/sepolicy
setools==4.4.4
setuptools==69.0.3
simpleaudio==1.0.4
simpleline==1.9.0
six==1.12.0
sortedcontainers==2.4.0
sos==4.7.0
soupsieve==2.5
SpeechRecognition==3.8.1
SQLAlchemy==2.0.31
systemd-python==235
Tempita==0.5.2
textract==1.6.5
typing_extensions==4.12.2
tzlocal==5.2
urllib3==1.26.18
Werkzeug==3.0.3
xkbregistry==0.3
xlrd==1.2.0
XlsxWriter==3.2.0
yarl==1.9.4
